Use primary transaction for useTransactionDisplayData
Fallback to initialTransaction if primaryTransaction is not present.

This is suppose to mainly fix how we present the subsequnt transaction of a failed transaction.
Since the failed transaction has the same nonce as the next proposed transaction, the data passed to the transaction by nonce will be the initial transaction data which is not true.
Ideally, we would always use the primaryTransaction data as the latest txCategory, value, from, to, etc.